charon-systemd: handle SIGUSR1 as error shutdown
To handle fatal errors of the kernel-danos plugins, there is need
of the plugin signal an error which results in a charon-systemd
shutdown which lets charon-systemd terminate with a non-zero return code.

A non-zero return code is required to leverage the systemd service
file Restart=on-abnormal functionality.

The SIGTERM signnal would let charon-systemd terminate with a return
code of zero. And not result in a strongswan restart.

SIGUSR1 is also used for error conditions in the charon-cmd
implementation (which is not used in DANOS).

VRVDR-54515